registration added

+++ src/main/java/com/iwkms/personalBlog/service/PostService.java
@@ -1,5 +1,6 @@
 package com.iwkms.personalBlog.service;
 import com.iwkms.personalBlog.model.Post;
+import com.iwkms.personalBlog.model.User;
 import com.iwkms.personalBlog.repository.PostRepository;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -29,25 +30,52 @@ public class PostService {
 
     @Transactional
     public Post createPost(Post post) {
+        if (post.getAuthor() == null) {
+            throw new IllegalArgumentException("Author cannot be null");
+        }
         return postRepository.save(post);
     }
 
     @Transactional
-    public Optional<Post> updatePost(Long id, Post postDetails) {
+    public Optional<Post> updatePost(Long id, Post postDetails, User currentUser) {
         return postRepository.findById(id)
-                .map(postEntity -> {
-                    postEntity.setTitle(postDetails.getTitle());
-                    postEntity.setContent(postDetails.getContent());
-                    return postRepository.save(postEntity);
+                .map(existingPost -> {
+                    if (!isAuthorizedToModify(existingPost, currentUser)) {
+                        throw new UnauthorizedAccessException("You are not authorized to update this post");
+                    }
+                    existingPost.setTitle(postDetails.getTitle());
+                    existingPost.setContent(postDetails.getContent());
+                    return postRepository.save(existingPost);
                 });
     }
 
     @Transactional
-    public void deletePost(Long id) {
-        try {
-            postRepository.deleteById(id);
-        } catch (Exception e) {
+    public void deletePost(Long id, User currentUser) {
+        Optional<Post> postToDelete = postRepository.findById(id);
+        if (postToDelete.isEmpty()) {
             throw new RuntimeException("Post with id " + id + " not found");
+        }
+        
+        if (!isAuthorizedToModify(postToDelete.get(), currentUser)) {
+            throw new UnauthorizedAccessException("You are not authorized to delete this post");
+        }
+        
+        postRepository.deleteById(id);
+    }
+    
+    private boolean isAuthorizedToModify(Post post, User user) {
+        // Admin can modify any post
+        if (user.getRoles().contains("ROLE_ADMIN")) {
+            return true;
+        }
+        
+        // Users can only modify their own posts
+        return post.getAuthor() != null && post.getAuthor().getId().equals(user.getId());
+    }
+    
+    public static class UnauthorizedAccessException extends RuntimeException {
+        public UnauthorizedAccessException(String message) {
+            super(message);
         }
     }
 }
